+++ RELEASE_NOTES.md
@@ -1,4 +1,13 @@
-﻿## 1.1.1
+﻿## 1.2.2
+ - QoS & PL settings can now be overriden if no endpoint configuration provided.
+
+## 1.2.1
+ - QoS & PL default settings are now preserved if not specified in configufation.
+
+## 1.2.0
+ - Incoming message listener QoS and parallelism level configuration added.
+
+## 1.1.1
  - Deleted destructor and overloaded Dispose method to leave only one option for endpoint deletion.
 
 ## 1.1.0